Add DisplayMapListCommandHandler and update Plugin class
Added a new class `DisplayMapListCommandHandler` in the `cs2_rockthevote.Features` namespace. This class implements the `IPluginDependency<Plugin, Config>` interface and handles the display of a map list from "maplist.txt". It also provides a "maplist" command for players to display a portion of the map list in the console.

Updated the `Plugin` class in the `cs2_rockthevote` namespace to include a new private readonly field `_displayMapListCommandHandler` of type `DisplayMapListCommandHandler`. The `Plugin` class constructor now takes an additional parameter of type `DisplayMapListCommandHandler`, allowing the `Plugin` class to utilize the functionality provided by the `DisplayMapListCommandHandler` class.